Added more weighted elements
