+++ composer.json
@@ -9,6 +9,7 @@
         "docs": "https://www.drupal.org/docs/user_guide/en/index.html"
     },
     "require": {
+        "php": "^8.3",
         "chx/drupal-issue-fork": "dev-main",
         "composer/installers": "^2.0",
         "cweagans/composer-patches": "^1.7",
