+++ config/sync/webform.webform.complexity_uncertainty_default.yml
@@ -13,6 +13,17 @@ title: 'Kan vi'
 description: ''
 categories: {  }
 elements: |-
+  name:
+    '#type': ai_screening_weighted_textfield
+    '#title': Name
+    '#description': 'Description of the name'
+    '#help_title': 'Help for the name'
+    '#ai_screening_weight_x': 1
+    '#ai_screening_weight_y': 2
+  description:
+    '#type': ai_screening_weighted_textarea
+    '#title': Description
+    '#description': "Describe what you're doing."
   radio_ga_ga:
     '#type': ai_screening_weighted_radios
     '#title': 'Radio Ga Ga'
@@ -23,7 +34,6 @@ elements: |-
       1: 'I -- The first option'
       2: 'II -- Another option'
       3: 'III -- The final option'
-    '#required': true
   radio_goo_goo:
     '#type': ai_screening_weighted_radios
     '#title': 'Radio Goo Goo'
@@ -33,7 +43,6 @@ elements: |-
       1: 'a -- The first option'
       2: 'b -- Another option'
       3: 'c -- The final option'
-    '#required': true
 css: ''
 javascript: ''
 settings:
