+++ web/modules/custom/ai_screening/src/Exception/Exception.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\ai_screening\Exception;
+
+/**
+ *
+ */
+class Exception extends \Exception {
+
+}
