+++ web/modules/custom/ai_screening/src/Exception/RuntimeException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\ai_screening\Exception;
+
+/**
+ *
+ */
+class RuntimeException extends Exception {
+
+}
