+++ web/modules/custom/ai_screening_project_track/ai_screening_project_track.info.yml
@@ -3,3 +3,5 @@ description: Adds project track entity and functionality
 package: 'Custom'
 type: module
 core_version_requirement: "^11"
+dependencies:
+  - ai_screening:ai_screening
