+++ web/modules/custom/ai_screening_project_track/src/Element/WeightedTextarea.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Drupal\ai_screening_project_track\Element;
+
+use Drupal\Core\Render\Element\Textarea;
+
+/**
+ * Weighted textarea element.
+ *
+ * @FormElement("ai_screening_weighted_textarea")
+ */
+class WeightedTextarea extends Textarea {
+
+}
