+++ web/modules/custom/ai_screening_project_track/src/Element/WeightedTextfield.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Drupal\ai_screening_project_track\Element;
+
+use Drupal\Core\Render\Element\Textfield;
+
+/**
+ * Weighted textfield element.
+ *
+ * @FormElement("ai_screening_weighted_textfield")
+ */
+class WeightedTextfield extends Textfield {
+
+}
