+++ web/modules/custom/ai_screening_project_track/src/Plugin/WebformElement/WeightedTextfield.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Drupal\ai_screening_project_track\Plugin\WebformElement;
+
+use Drupal\webform\Plugin\WebformElement\TextField;
+
+/**
+ * Weighted radios element.
+ *
+ * @WebformElement(
+ *   id = "ai_screening_weighted_textfield",
+ *   label = @Translation("Weighted textfield"),
+ *   description = @Translation("Texfield with an (x, y) weight."),
+ *   category = @Translation("AI Screening"),
+ * )
+ */
+class WeightedTextfield extends TextField {
+  use WeightedElementTrait;
+
+}
